\docType{package}
\name{VineCopulaHD-package}
\alias{VineCopulaHD-package}
\alias{VineCopulaHD}
\title{Several Methods To Estimate High Dimensional Vine Copulas}
\description{
Vine Copulas are very flexible statistical models and are basis of the package VineCopula. However, they become quite complex in higher dimensions (d > 50). This package
    uses methods from graphical models and univariate model selection to estimate high dimensional vine copula structures and sparsity patterns.
}
\author{Dominik M端ller}
\references{
M端ller, D. and C. Czado. (2017).
Representing Sparse Gaussian DAGs as Sparse R-vines Allowing for Non-Gaussian Dependence.
http://amstat.tandfonline.com/doi/abs/10.1080/10618600.2017.1366911

M端ller, D. and C. Czado. (2017).
Selection of Sparse Vine Copulas in High Dimensions with the Lasso.
https://arxiv.org/abs/1705.05877

M端ller, D. and C. Czado. (2017).
Dependence Modeling in Ultra High Dimensions with Vine Copulas and the Graphical Lasso.
https://arxiv.org/abs/1709.05119

Ulf Schepsmeier, Jakob Stoeber, Eike Christian Brechmann, Benedikt Graeler, Thomas Nagler and Tobias Erhardt (NA). VineCopula: Statistical Inference of Vine Copulas.
R package version 2.1.4.9000.
https://github.com/tnagler/VineCopula}
